+++ src/LinAlg/hiopMatrixMDS.hpp
@@ -45,7 +45,7 @@ public:
     mDe->copyFrom(*m.mDe);
   }
 
-  virtual void copyRowsFrom(const hiopMatrix& src_in, const int* rows_idxs, int n_rows)
+  virtual void copyRowsFrom(const hiopMatrix& src_in, const long long* rows_idxs, long long n_rows)
   {
     const hiopMatrixMDS& src = dynamic_cast<const hiopMatrixMDS&>(src_in);
     mSp->copyRowsFrom(src, rows_idxs, n_rows);
@@ -243,7 +243,7 @@ public:
     mDe->copyFrom(*m.mDe);
   }
 
-  virtual void copyRowsFrom(const hiopMatrix& src_in, const int* rows_idxs, int n_rows)
+  virtual void copyRowsFrom(const hiopMatrix& src_in, const long long* rows_idxs, long long n_rows)
   {
     const hiopMatrixSymBlockDiagMDS& src = dynamic_cast<const hiopMatrixSymBlockDiagMDS&>(src_in);
     mSp->copyRowsFrom(src, rows_idxs, n_rows);
